Use a mock implementation.
